backend/requirements.txt
apispec==6.3.0
asgiref==3.7.2
dataclasses==0.6
Django==4.1.9
django-cors-headers==4.3.1
django-serializer==1.2.1
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.4.2
marshmallow==3.20.1
packaging==23.2
pymongo==4.6.0
pytz==2023.3.post1
sqlparse==0.2.4
tzdata==2023.3
